rework: move utilities to utils.c
